[refactor] Move CompiledFieldData to aot namespace.
