+++ taichi/backends/vulkan/aot_utils.h
@@ -3,6 +3,7 @@
 #include <vector>
 
 #include "taichi/codegen/spirv/kernel_utils.h"
+#include "taichi/program/aot_module_builder.h"
 
 namespace taichi {
 namespace lang {
@@ -15,7 +16,7 @@ struct TaichiAotData {
   //   BufferMetaData metadata;
   std::vector<std::vector<std::vector<uint32_t>>> spirv_codes;
   std::vector<spirv::TaichiKernelAttributes> kernels;
-  std::vector<spirv::CompiledFieldData> fields;
+  std::vector<aot::CompiledFieldData> fields;
 
   TI_IO_DEF(kernels, fields);
 };
